scripts/requirements.txt
supabase==2.9.0
python-dotenv==1.0.0
